+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is inspired by [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [0.9.3] - 2026-02-05
+
+### Added
+
+- Inbox persistence support with `PersistencePolicy` and `InboxPersistence` options
+- `Persistent` property on `IInbox` and `InboxExport`
+- `PersistencePolicy` and `PersistentGlobalWebhooks` on `ServerInfo`
+
 ## [0.9.2] - 2026-01-29
 
 ### Changed